api/requirements.txt
alembic
bcrypt
fastapi
opencv-python-headless
passlib
psycopg2
pydantic
pydantic-settings
pydantic[email]
python-jose
python-multipart
sqlalchemy
ultralytics
uvicorn
